refactored create tests
